Minor push.
Console is colored.
When the swear filter is activated, it'll log what caused it to activate
in #censor-log
Added more checks on image scanning, in case the scan fails.
Moderators are now immune to the swear filter.
To use .ping, you now need to be a moderator. This change is in direct
response to spam of the command.
Added a .sort command to remove any duplicates of a word from the swear
filter on a specific server. Only usable by me, because it can lag the
bot pretty badly.